Finished swap chain implementation Moved most AppInstance member variables outside. Reorganized classes.

+++ utils/Window.cpp
@@ -1,11 +1,9 @@
 #include "Window.h"
 
-#include <cassert>
-
 #include "DebugUtils.h"
 
 namespace vk {
-Window::Window(const int width, const int height, const char* title) {
+Window::Window(const uint32_t width, const uint32_t height, const char* title) {
 	assert(width > 0);
 	assert(height > 0);
 	assert(title != nullptr);
@@ -18,7 +16,7 @@ Window::Window(const int width, const int height, const char* title) {
 	// Avoid window resizing
 	glfwWindowHint(GLFW_RESIZABLE, GLFW_FALSE);
 
-	mWindow = glfwCreateWindow(width, height, title, nullptr, nullptr);
+	mWindow = glfwCreateWindow(static_cast<int>(width), static_cast<int>(height), title, nullptr, nullptr);
 	assert(mWindow != nullptr && "Window creation failed");
 }
